Final Merged/requirements.txt
flask
requests
numpy
opencv-python
face_recognition
picamera2
pyserial
